Create PingIt
Powerful PowerShell Script that will check for devices online. It does this by going through a list of delimited devices. If a device is found (returns ping request successfully) it will then either run a cmd or ps script based on the toggles set to perform admin actions on device. if found it will remove and log the actions. If not it will loop again and reattempt for 'x' amount of time. I will attempt list 'x' times and then close. If all items completed, then it will close session.